modify deleteAccount endpoint

+++ Project/Backend/src/modules/user/services/deleteAccount.service.js
@@ -1,6 +1,16 @@
 import User from "../../../DB/model/User.model.js"
+import bcrypt from 'bcrypt'
 const deleteAccount=async(req , res , next)=>{
     try {
+        const {oldPassword}=req.body;
+        const checkUser=await User.findById(req.params.id);
+        console.log(checkUser.password);
+        const match=bcrypt.compareSync(oldPassword , checkUser.password)
+        if(!match){
+            return res.status(404).json({message: 'incorrect password'})
+        }
+        console.log("matched");
+        
         const deletedUser=await User.findByIdAndDelete(
             req.params.id,
             { new: true }
